Header fix
We need to create a container for the image, the hr, and the headers. It will ease the formatting and allow us to effectively use ratios. I should be able to figure this out. We should ideally be using a container for everything because css is a lot of sizing and formatting and it's so much easier with containers.